# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "EvacuationPointRequest": {
        "properties": {
          "capacity_households": {
            "exclusiveMinimum": 0.0,
            "maximum": 100000.0,
            "title": "Capacity Households",
            "type": "integer"
          },
          "contact_name": {
            "maxLength": 180,
            "minLength": 2,
            "title": "Contact Name",
            "type": "string"
          },
          "contact_phone": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Phone"
          },
          "latitude": {
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude",
            "type": "number"
          },
          "longitude": {
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude",
            "type": "number"
          },
          "name": {
            "maxLength": 180,
            "minLength": 3,
            "title": "Name",
            "type": "string"
          },
          "village_id": {
            "format": "uuid",
            "title": "Village Id",
            "type": "string"
          }
        },
        "required": [
          "village_id",
          "name",
          "latitude",
          "longitude",
          "capacity_households",
          "contact_name"
        ],
        "title": "EvacuationPointRequest",
        "type": "object"
      },
      "EvacuationPointVerificationRequest": {
        "properties": {
          "is_verified": {
            "title": "Is Verified",
            "type": "boolean"
          }
        },
        "required": [
          "is_verified"
        ],
        "title": "EvacuationPointVerificationRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Ba Na SmartLink API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/pilots/alerts": {
      "get": {
        "operationId": "list_internal_alerts_api_pilots_alerts_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Internal Alerts Api Pilots Alerts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Internal Alerts",
        "tags": [
          "feature-flagged-pilots"
        ]
      }
    },
    "/api/pilots/evacuation-points": {
      "get": {
        "description": "Return verified evacuation points without private contact details.\n\nThis is a public preparedness directory, not an emergency alert channel.",
        "operationId": "list_public_evacuation_points_api_pilots_evacuation_points_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Public Evacuation Points Api Pilots Evacuation Points Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Public Evacuation Points",
        "tags": [
          "feature-flagged-pilots"
        ]
      },
      "post": {
        "description": "Create an unverified point; publication requires a separate admin action.",
        "operationId": "create_evacuation_point_api_pilots_evacuation_points_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvacuationPointRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Evacuation Point Api Pilots Evacuation Points Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Evacuation Point",
        "tags": [
          "feature-flagged-pilots"
        ]
      }
    },
    "/api/pilots/evacuation-points/admin": {
      "get": {
        "description": "Return the internal evacuation catalogue for admin/leader review.",
        "operationId": "list_admin_evacuation_points_api_pilots_evacuation_points_admin_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Admin Evacuation Points Api Pilots Evacuation Points Admin Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Admin Evacuation Points",
        "tags": [
          "feature-flagged-pilots"
        ]
      }
    },
    "/api/pilots/evacuation-points/{point_id}/verification": {
      "patch": {
        "description": "Publish or withdraw a point after the admin has verified its source.",
        "operationId": "verify_evacuation_point_api_pilots_evacuation_points__point_id__verification_patch",
        "parameters": [
          {
            "in": "path",
            "name": "point_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Point Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvacuationPointVerificationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Verify Evacuation Point Api Pilots Evacuation Points  Point Id  Verification Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Evacuation Point",
        "tags": [
          "feature-flagged-pilots"
        ]
      }
    },
    "/api/pilots/sensors/devices": {
      "get": {
        "operationId": "list_sensor_devices_api_pilots_sensors_devices_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Sensor Devices Api Pilots Sensors Devices Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sensor Devices",
        "tags": [
          "feature-flagged-pilots"
        ]
      },
      "post": {
        "operationId": "create_sensor_device_api_pilots_sensors_devices_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorDeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Sensor Device Api Pilots Sensors Devices Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Sensor Device",
        "tags": [
          "feature-flagged-pilots"
        ]
      }
    },
    "/api/pilots/sensors/observations": {
      "get": {
        "description": "Return recent pilot observations for internal review only.\n\nThis endpoint deliberately exposes raw readings only to admin/leader and never\ncreates alerts or public messages by itself.",
        "operationId": "list_sensor_observations_api_pilots_sensors_observations_get",
        "parameters": [
          {
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Sensor Observations Api Pilots Sensors Observations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sensor Observations",
        "tags": [
          "feature-flagged-pilots"
        ]
      },
      "post": {
        "operationId": "ingest_sensor_observation_api_pilots_sensors_observations_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SensorObservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ingest Sensor Observation Api Pilots Sensors Observations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Sensor Observation",
        "tags": [
          "feature-flagged-pilots"
        ]
      }
    },
    "/api/pilots/status": {
      "get": {
        "description": "Expose feature flags to the internal pilot workbench without exposing data.",
        "operationId": "pilot_status_api_pilots_status_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Pilot Status Api Pilots Status Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Pilot Status",
        "tags": [
          "feature-flagged-pilots"
        ]
      }
    },
    "/api/pilots/tourism/places": {
      "get": {
        "operationId": "list_public_tourism_places_api_pilots_tourism_places_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": true,
                    "type": "object"
                  },
                  "title": "Response List Public Tourism Places Api Pilots Tourism Places Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Public Tourism Places",
        "tags": [
          "feature-flagged-pilots"
        ]
      },
      "post": {
        "operationId": "create_tourism_place_api_pilots_tourism_places_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TourismPlaceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Tourism Place Api Pilots Tourism Places Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tourism Place",
        "tags": [
          "feature-flagged-pilots"
        ]
      }
    },
...
  }
}